사용자 권한 작업

+++ restAPI/controllers/feed.js
@@ -1,36 +1,37 @@
 const fs = require('fs');
 const path = require('path');
 
-const { validationResult, Result } = require('express-validator')
+const { validationResult } = require('express-validator')
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
   const perPage = 2;
   let totalItems;
   Post.find()
-  .countDocuments()
-  .then(count => {
-    totalItems = count;
-    return Post.find()
-    .skip( (currentPage - 1 ) * perPage)
-    .limit(perPage);
-  })
-  .then(posts => {
-    res.status(200)
-    .json({ 
-      message: 'Fetched posts successfully', 
-      posts: posts, 
-      totalItems: totalItems 
+    .countDocuments()
+    .then(count => {
+      totalItems = count;
+      return Post.find()
+        .skip((currentPage - 1) * perPage)
+        .limit(perPage);
     })
-  })
-  .catch(err => {
-    if (!err.statusCode) {
-      err.statusCode = 500;
-    }
-    next(err);
-  });
+    .then(posts => {
+      res.status(200)
+        .json({
+          message: 'Fetched posts successfully',
+          posts: posts,
+          totalItems: totalItems
+        })
+    })
+    .catch(err => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
 };
 
 exports.createPost = (req, res, next) => {
@@ -48,18 +49,30 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path.replace("\\", "/");
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'ysj' },
+    creator: req.userId,
   })
   post.save()
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: {
+          _id: creator._id,
+          name: creator.name
+        }
       });
     })
     .catch(err => {
@@ -104,7 +117,7 @@ exports.updatedpost = (req, res, next) => {
     imageUrl = req.file.path.replace("\\", "/");
   }
   if (!imageUrl) {
-    error = new Error('No file picked');
+    const error = new Error('No file picked');
     error.statusCode = 422;
     throw error;
   }
@@ -113,6 +126,11 @@ exports.updatedpost = (req, res, next) => {
       if (!post) {
         const error = new Error('Could not find post');
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -137,25 +155,36 @@ exports.updatedpost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
-  .then(post => {
-    if (!post) {
-      const error = new Error('Could not find post');
-      error.statusCode = 404;
-      throw error;
-    }
-    clearImage(post.imageUrl);
-    return Post.findByIdAndRemove(postId);
-  })
-  .then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Deleted post.'});
-  })
-  .catch(err => {
-    if (!err.statusCode) {
-      err.statusCode = 500;
-    }
-    next(err);
-  })
+    .then(post => {
+      if (!post) {
+        const error = new Error('Could not find post');
+        error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
+      clearImage(post.imageUrl);
+      return Post.findByIdAndRemove(postId);
+    })
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: 'Deleted post.' });
+    })
+    .catch(err => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    })
 }
 
 const clearImage = filePath => {
